requirements.txt
flask
sqlalchemy
python-dotenv
openai>=1.30.0
Jinja2==3.1.4
requests>=2.32.0
python-dotenv==1.0.1
gunicorn==22.0.0